Serialization now works on FastStringBuilder, which means we can archive and load the vast majority of strings in use by a script.
Next step: when a new object ref is saved that is not a string type, we map it to a string refernce, the name of the, and save the file in the archive.

When we load the object we check the objects we have so far created and set their members whence an object-ref was supplied for defered assignment.