We can now load and save arrays of interfaces. We also fixed a bug where an array of interfaces had the incorrect element size(something other than 8).
We need to ensure we can load and save interfaces that are part of structures that are the array elements.

We need to refactor the archive code so that objects list a block of fields then  a block of values, rather than combine both fields and values on the same line. This will allow us to re-use the serialization code and reduce the code bloat.

Once refactoring is done we should move on to the serialization of maps and linked lists.