Improvements to the components builder. Fixed duplicate native calls registration throughout the codebase.
Introduced unified memory allocation in Sexy, with the creation of SexyAllocator. There are still some items that need updating to the new system, but it is still far better than it was.

The new allocator was used to identify major memory leaks which appear to all have been fixed, including those associated with duplicate native call registration, which now throws an error.

TODO -> Use some kind of typedef, macro or using directive to reduce declaration size.

TODO -> Use the SexyAllocator with the fast stringmap.

TODO -> Use the SexyAllocator with profiling to improve compilation speed and reduce allocation costs.