Added better error reporting for bugs discovered in native files when the script system is constructed.
Added (Sys.Type.string <builder-name> <capacity> args...) for allocating string  builders and appending arguments in one line. The test is TestMacroSiblings2, which does not work, as the outIndex, computed at 5, magically becomes 0. We must debug in sexy.dotnet, but it appears broken since the last VS update. The tests is disabled for now.