According to docs, fullscreen apps should have zero for classDef.hbrBackground. This was done in the dx11.factory.cpp file.
Also according to the docs exiting the app should first switch to window mode, else dx11 resources are not released. This was consistent with the debug output. The correction is in place.

Full screen mode switch now stored the fullscreen dimenisions in the windowbacking object. Previously switching to fullscreen from window mode reset to the default resolution of the system.