Added sexy.domme (D)ata (O)bject (M)odel (M)egic (E)gg
We already have a system that exposes C++ objects to Sexy, but we want to expose a Sexy object to C++. I figured out how to do it - run a function that passes a struct arg that has an interface member. In the function assign a new object to the interface and then yield. Since the yield preserves the callstack the object will not be freed and the caller can take the interface from the stack and invoke the methods on it. This allows a persistent script to expose its methods to C++